Create try_except
Learning about the try and except block, similar to the if else statement. Used when a section of code may produce an error or an exception for example when getting data from a website (the server might be down) or getting data from an API (the API might no produce the excepted result).